+++ src/django_components/__init__.py
@@ -53,6 +53,7 @@ from django_components.tag_formatter import (
 from django_components.template import cached_template
 import django_components.types as types
 from django_components.util.loader import ComponentFileEntry, get_component_dirs, get_component_files
+from django_components.util.routing import URLRoute, URLRouteHandler
 from django_components.util.types import EmptyTuple, EmptyDict
 
 # isort: on
@@ -113,4 +114,6 @@ __all__ = [
     "TagResult",
     "template_tag",
     "types",
+    "URLRoute",
+    "URLRouteHandler",
 ]