+++ src/django_components/util/command.py
@@ -228,7 +228,7 @@ class ComponentCommand:
     For example, if you define and install the following extension:
 
     ```python
-    from django_components ComponentCommand, ComponentsExtension
+    from django_components ComponentCommand, ComponentExtension
 
     class HelloCommand(ComponentCommand):
         name = "hello"
@@ -236,7 +236,7 @@ class ComponentCommand:
         def handle(self, *args, **kwargs):
             print("Hello, world!")
 
-    class MyExt(ComponentsExtension):
+    class MyExt(ComponentExtension):
         name = "my_ext"
         commands = [HelloCommand]
     ```
@@ -250,7 +250,7 @@ class ComponentCommand:
     You can also define arguments for the command, which will be passed to the command's `handle` method.
 
     ```python
-    from django_components import CommandArg, ComponentCommand, ComponentsExtension
+    from django_components import CommandArg, ComponentCommand, ComponentExtension
 
     class HelloCommand(ComponentCommand):
         name = "hello"